Change fields_struct::offset to size_t from uint32.
Former-commit-id: 060f7001e0b02cab2d14641d20fd6c30a9897f14

+++ server/src/plugins/plugin_python/include/plugin_python.h
@@ -383,7 +383,7 @@ typedef struct
 
 	/**
 	 * Offset in player structure. */
-	uint32 offset;
+	size_t offset;
 
 	/**
 	 * Flags for special handling. */
